feat: update create_room
Changed how rooms will be created, so when a `create_room` event is sent,
we don't need the game name. Only when the game starts will then a game be
set for a room. This leads to a better UX, as the players can change the game
of the room right until they play. Saves creating a new room etc.

Also updated room codes to be more than 5 character all upper case. Can now
be [A-Za-z0-9] of length 12. As I expect the user to copy the invite link.
Rather than manually enter a room code.

Fixes: #20, #21